requirements.txt
streamlit
scikit-learn
xgboost
lightgbm
joblib
numpy
pandas
python-docx
PyPDF2
pdfplumber
